changed function call in vm translator
When running the NestedCall tests, all output RAM values would
be incorrect.  This was being caused by a "stack skew".  The
function call was pushing N things to the stack, and the
function return was NOT popping N things!

Possible culprit: the function call was previously NOT pushing
the "THAT" pointer to the stack, which is probably why errors
were propagating throughout any programs which made use of it.

This also would explain why the errors don't always show up.
Any tests which do not use "THAT" would probably have passed,
and any tests which use it would have failed spectacularly.