First Parser Test has passed!  Still more to go.
The testing package has greatly improved the ability to debug
this program.  Since it will install the go program, and then
externally call the code, it gives a more "realistic" test
for how the program will be used.

The most useful aspect is the ability to quickly see at which
line the output differs from the expected output.  This allows
errors to be pinpointed and debugged quickly.

Found 5+ small bugs in the parser (several were the exact
same type of mistake, but were found in different parts of the
code).

	hackcompiler Main.jack -p > Main.xml

This program has been succesfully compiled and compared to its
expected output!  I imagine there are only a few more bugs to
work out before the syntatic anlyzer is complete!